LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/logical-overflow-001-expected.txt


FAIL overflow-x matches overflow-inline, and overflow-y matches overflow-block when the element has a horizontal writing mode assert_equals: expected "scroll" but got "visible"
FAIL overflow-y matches overflow-inline, and overflow-x matches overflow-block when the element has a vertical writing mode assert_equals: expected "hidden" but got "visible"
